:sparkles: feat: Created the StaticHeader and stories
