+++ src/components/header/header.mdx
@@ -1,0 +1,19 @@
+import { Meta, Controls, Primary, Canvas } from '@storybook/blocks'
+import * as Story from './header.stories'
+
+<Meta of={Story} />
+
+# Header
+
+The main header component for the application.
+
+<Canvas of={Story.Primary} />
+
+### Static Header
+
+An static version, used on the onboarding user flow.
+
+<Canvas of={Story.StaticHeaderExample} />
+
+<Primary />
+<Controls />
